+++ ghostwriter/reporting/views2/report.py
@@ -29,7 +29,7 @@ from django.contrib import messages
 from django.utils import dateformat, timezone
 from django.utils.html import strip_tags
 from channels.layers import get_channel_layer
-from ghostwriter.api.utils import RoleBasedAccessControlMixin, get_reports_list, get_templates_list, verify_user_is_privileged
+from ghostwriter.api.utils import RoleBasedAccessControlMixin, get_reports_list, get_templates_list
 from ghostwriter.commandcenter.models import BloodHoundConfiguration, ExtraFieldSpec, ReportConfiguration
 from ghostwriter.commandcenter.views import CollabModelUpdate, ExtraFieldJsonView, ExtraFieldRichTextPreviewView
 from ghostwriter.modules.exceptions import MissingTemplate
@@ -818,17 +818,14 @@ class ReportTemplateUpdate(RoleBasedAccessControlMixin, UpdateView):
     template_name = "reporting/report_template_form.html"
 
     def test_func(self):
-        obj = self.get_object()
-        if obj.protected:
-            return verify_user_is_privileged(self.request.user)
-        return obj.user_can_view(self.request.user)
+        return self.get_object().user_can_edit(self.request.user)
 
     def handle_no_permission(self):
         obj = self.get_object()
-        if obj.protected:
+        if obj.protected or obj.client_id is None:
             messages.error(
                 self.request,
-                "That template is protected – only admins and managers can edit it.",
+                "Report template management permission is required to edit protected or global templates.",
             )
         else:
             messages.error(self.request, "You do not have permission to access that.")
@@ -892,7 +889,7 @@ class ReportTemplateDelete(RoleBasedAccessControlMixin, DeleteView):
     template_name = "confirm_delete.html"
 
     def test_func(self):
-        return verify_user_is_privileged(self.request.user)
+        return self.get_object().user_can_delete(self.request.user)
 
     def handle_no_permission(self):
         messages.error(self.request, "You do not have permission to access that.")